app/requirements.txt
Flask==3.0.0
Flask-Mail==0.10.0
mysql-connector-python
